is_mate bugfix
Fixed bug with is_mate where a square that a pawn was attacking could
block a checking piece. The pawn couldn't actually move to this square,
so a fix was implemented that checks if the pawn can actually move to
the square where it is trying to block mate.

Additionally tried to add heatmaps for piece advantageous squares. After
implementing a new way to sort the moves in order_moves I tested it, and
it was far slower than the old way. Just commiting for documentation.