Fixed bugs
Fixed a bug where the cycle counter (struct version not DWT) would
frequently reset due to a u32 value being used in a multiplication
instead of a u64. This would cause the game to freeze because the time
would reset while the minimax algorithm is running.

Additionally fixed the computer telling you to castle a turn after it
actually castled.