Added OpenSCAD models to design
Added OpenSCAD models for spacers that are used for mounting the
character lcd, and plate ontop of the PCB. These were initially intended
to be washers, but 3D printing allows them to be one piece rather than
multiple stacked washers. This makes assembly easier. Additionally added
an OpenSCAD file for a chess base piece, which can be used to convert
normal chess pieces into ones with magnets.